ec2-report/requirements.txt
aws-cdk-lib>=2.27.0
constructs>=10.0.0
pytest==6.2.5
boto3
